[refactor] website: add react HOC for Modal
