+++ website/src/App.js
@@ -6,9 +6,15 @@ import { config, defaultDataValue, defaultSettings, themeColor } from './constan
 
 import NavBar from './components/NavBar';
 import TextArea from './components/TextArea';
-import ConnectingModal from './components/ConnectingModal';
-import NewRoomJoinModal from './components/NewRoomJoinModal';
-import SettingsModal from './components/SettingsModal';
+import ModalWrapper from './components/ModalWrapper';
+
+import ConnectingImg from './components/Connecting';
+import RoomInfo from './components/RoomInfo';
+import Settings from './components/Settings';
+
+const SettingsModal = ModalWrapper(Settings);
+const ConnectingModal = ModalWrapper(ConnectingImg);
+const RoomInfoModal = ModalWrapper(RoomInfo);
 
 const App = () => {
     const [data, setData] = useState(defaultDataValue);
@@ -56,10 +62,6 @@ const App = () => {
 
     const copyAllText = () => navigator.clipboard.writeText(data.value);
 
-    const openSettings = () => setCurrentModal('Settings');
-
-    const openNewSessionJoin = () => setCurrentModal('NewSessionJoin');
-
     const closeModal = () => setCurrentModal('');
 
     const handleResponse = response => {
@@ -101,8 +103,8 @@ const App = () => {
             <NavBar
                 usersCount={usersCount}
                 roomId={roomId}
-                openNewSessionJoin={openNewSessionJoin}
-                openSettings={openSettings}
+                openRoomInfo={() => setCurrentModal('RoomInfo')}
+                openSettings={() => setCurrentModal('Settings')}
                 copyAllText={copyAllText}
                 cutAllText={cutAllText}
             />
@@ -110,7 +112,7 @@ const App = () => {
             {currentModal === '' && (
                 <TextArea value={data.value} handleChange={handleChange} fontSize={settings.fontSize} themeColor={themeColor[settings.theme]} />
             )}
-            {currentModal === 'NewSessionJoin' && <NewRoomJoinModal handleRoomJoin={handleRoomJoin} closeModal={closeModal} roomId={roomId} />}
+            {currentModal === 'RoomInfo' && <RoomInfoModal handleRoomJoin={handleRoomJoin} closeModal={closeModal} roomId={roomId} />}
             {currentModal === 'Settings' && <SettingsModal closeModal={closeModal} settings={settings} setSettings={setSettings} />}
         </div>
     );
